well, I have 80% test coverage and pylint >7. I can run the tests locally on the command line or my PyCharm IDE and I can also debug them in PyCharm. That feels like a win, so I'm going to push what I have so far.
I'm struggling right now to come up with a CI/CD solution.  I started with Github Action workflows but I wanted to run in AWS CodeBuild to make it easy to integration into an AWS Pipeline for deployment.   I ran into problems because my integration tests use docker compose to set up  db and mail servers as containers and testing the workflow using act fails because the networking gets tricky.

I'm going to pause getting the CI/CD setup working and do some more research.   What I want is to have the tests run automatically when I push to main and I want to use as vendor-neutral a setup as possible to define the workflow.